Create README.md
The program calculates an employee or an employer's salary based  on working hours in a month .This program assumes the total number of working hours is 20 hours. The program also assumes an employee is paid $160.00 and employer is paid $1000.00 each hour he or she attends work. The program also assumes the employee or the employer enjoys 10% benefit on the salary if the person did not absent him or herself for a day
The program also assumes an employer is not entitled to any ddeduction.